add getActiveConfiguration within OpfPaymentFacade
