+++ integration-libs/opf/base/core/connectors/opf-payment.adapter.ts
@@ -5,6 +5,7 @@
  */
 
 import {
+  ActiveConfiguration,
   AfterRedirectScriptResponse,
   OpfPaymentVerificationPayload,
   OpfPaymentVerificationResponse,
@@ -48,4 +49,9 @@ export abstract class OpfPaymentAdapter {
   abstract afterRedirectScripts(
     paymentSessionId: string
   ): Observable<AfterRedirectScriptResponse>;
+
+  /**
+   * Abstract method used to get payment active configurations
+   */
+  abstract getActiveConfigurations(): Observable<ActiveConfiguration[]>;
 }
