+++ integration-libs/opf/base/core/facade/facade-providers.ts
@@ -6,7 +6,6 @@
 
 import { Provider } from '@angular/core';
 import {
-  OpfConfigurationFacade,
   OpfGlobalFunctionsFacade,
   OpfOrderFacade,
   OpfOtpFacade,
@@ -14,7 +13,6 @@ import {
 } from '@spartacus/opf/base/root';
 
 import { OpfPaymentHostedFieldsService } from '../services/opf-payment-hosted-fields.service';
-import { OpfConfigurationService } from './opf-configuration.service';
 import { OpfGlobalFunctionsService } from './opf-global-functions.service';
 import { OpfOrderService } from './opf-order.service';
 import { OpfOtpService } from './opf-otp.service';
@@ -26,7 +24,6 @@ export const facadeProviders: Provider[] = [
   OpfOrderService,
   OpfOtpService,
   OpfGlobalFunctionsService,
-  OpfConfigurationService,
   {
     provide: OpfPaymentFacade,
     useExisting: OpfPaymentService,
@@ -42,9 +39,5 @@ export const facadeProviders: Provider[] = [
   {
     provide: OpfGlobalFunctionsFacade,
     useExisting: OpfGlobalFunctionsService,
-  },
-  {
-    provide: OpfConfigurationFacade,
-    useExisting: OpfConfigurationService,
   },
 ];
