+++ integration-libs/opf/base/core/facade/index.ts
@@ -4,9 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './opf-configuration.service';
 export * from './opf-global-functions.service';
 export * from './opf-order.service';
 export * from './opf-otp.service';
 export * from './opf-payment.service';
-
