+++ integration-libs/opf/base/core/facade/opf-payment.service.ts
@@ -5,8 +5,15 @@
  */
 
 import { Injectable } from '@angular/core';
-import { Command, CommandService } from '@spartacus/core';
 import {
+  Command,
+  CommandService,
+  Query,
+  QueryService,
+  QueryState,
+} from '@spartacus/core';
+import {
+  ActiveConfiguration,
   AfterRedirectScriptResponse,
   OpfPaymentFacade,
   OpfPaymentVerificationPayload,
@@ -66,7 +73,13 @@ export class OpfPaymentService implements OpfPaymentFacade {
     );
   });
 
+  protected activeConfigurationsQuery: Query<ActiveConfiguration[]> =
+    this.queryService.create<ActiveConfiguration[]>(() =>
+      this.opfPaymentConnector.getActiveConfigurations()
+    );
+
   constructor(
+    protected queryService: QueryService,
     protected commandService: CommandService,
     protected opfPaymentConnector: OpfPaymentConnector,
     protected opfPaymentHostedFieldsService: OpfPaymentHostedFieldsService
@@ -96,5 +109,11 @@ export class OpfPaymentService implements OpfPaymentFacade {
 
   afterRedirectScripts(paymentSessionId: string) {
     return this.afterRedirectScriptsCommand.execute({ paymentSessionId });
+  }
+
+  getActiveConfigurationsState(): Observable<
+    QueryState<ActiveConfiguration[] | undefined>
+  > {
+    return this.activeConfigurationsQuery.getState();
   }
 }
