+++ integration-libs/opf/base/occ/adapters/index.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './occ-opf-configuration.adapter';
 export * from './occ-opf-order.adapter';
 export * from './occ-opf.adapter';
 export * from './occ-otp.adapter';
