--- integration-libs/opf/base/occ/adapters/occ-opf-configuration.adapter.ts
@@ -1,60 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import {
-  backOff,
-  ConverterService,
-  isJaloError,
-  normalizeHttpError,
-} from '@spartacus/core';
-import {
-  OpfConfigurationAdapter,
-  OpfEndpointsService,
-  OPF_ACTIVE_CONFIGURATION_NORMALIZER,
-} from '@spartacus/opf/base/core';
-import {
-  ActiveConfiguration,
-  OpfConfig,
-  OPF_CC_PUBLIC_KEY,
-} from '@spartacus/opf/base/root';
-
-import { Observable, throwError } from 'rxjs';
-import { catchError } from 'rxjs/operators';
-
-@Injectable()
-export class OccOpfConfigurationAdapter implements OpfConfigurationAdapter {
-  constructor(
-    protected http: HttpClient,
-    protected converter: ConverterService,
-    protected opfEndpointsService: OpfEndpointsService,
-    protected config: OpfConfig
-  ) {}
-
-  getActiveConfigurations(): Observable<ActiveConfiguration[]> {
-    const headers = new HttpHeaders().set(
-      OPF_CC_PUBLIC_KEY,
-      this.config.opf?.commerceCloudPublicKey || ''
-    );
-
-    return this.http
-      .get<ActiveConfiguration[]>(this.getActiveConfigurationsEndpoint(), {
-        headers,
-      })
-      .pipe(
-        catchError((error) => throwError(normalizeHttpError(error))),
-        backOff({
-          shouldRetry: isJaloError,
-        }),
-        this.converter.pipeable(OPF_ACTIVE_CONFIGURATION_NORMALIZER)
-      );
-  }
-
-  protected getActiveConfigurationsEndpoint(): string {
-    return this.opfEndpointsService.buildUrl('getActiveConfigurations');
-  }
-}
