+++ integration-libs/opf/base/occ/config/default-occ-opf-config.ts
@@ -15,6 +15,7 @@ export const defaultOccOpfConfig: OccConfig = {
         submitCompletePayment: 'payments/${paymentSessionId}/submit-complete',
         afterRedirectScripts:
           'payments/${paymentSessionId}/after-redirect-scripts',
+        getActiveConfigurations: 'active-configurations',
       },
     },
   },
