--- integration-libs/opf/base/occ/config/default-occ-opf-configuration-config.ts
@@ -1,17 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { OccConfig } from '@spartacus/core';
-
-export const defaultOccOpfConfigurationConfig: OccConfig = {
-  backend: {
-    occ: {
-      endpoints: {
-        getActiveConfigurations: 'active-configurations',
-      },
-    },
-  },
-};
