+++ integration-libs/opf/base/occ/model/index.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './occ-opf-configuration-endpoints.model';
 export * from './occ-opf-endpoints.model';
 export * from './occ-opf-order-enpoints.model';
 export * from './occ-otp-endpoints.model';
