+++ integration-libs/opf/base/occ/model/occ-opf-endpoints.model.ts
@@ -25,5 +25,9 @@ declare module '@spartacus/core' {
      * Endpoint to fetch dynamic script for Hosted Fields pattern and PageRedirection sub-pattern.
      */
     afterRedirectScripts?: string | OccEndpoint;
+    /**
+     * Endpoint to get active payment configurations
+     */
+    getActiveConfigurations?: string | OccEndpoint;
   }
 }
