+++ integration-libs/opf/base/occ/opf-base-occ.module.ts
@@ -8,21 +8,15 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { provideDefaultConfig } from '@spartacus/core';
 import {
-  OpfConfigurationAdapter,
   OpfOrderAdapter,
   OpfPaymentAdapter,
   OtpAdapter,
 } from '@spartacus/opf/base/core';
 import { OccOrderNormalizer } from '@spartacus/order/occ';
 import { ORDER_NORMALIZER } from '@spartacus/order/root';
-import {
-  OccOpfConfigurationAdapter,
-  OccOpfOrderAdapter,
-  OccOtpAdapter,
-} from './adapters';
+import { OccOpfOrderAdapter, OccOtpAdapter } from './adapters';
 import { OccOpfPaymentAdapter } from './adapters/occ-opf.adapter';
 import { defaultOccOpfConfig } from './config/default-occ-opf-config';
-import { defaultOccOpfConfigurationConfig } from './config/default-occ-opf-configuration-config';
 import { defaultOccOpfOrderConfig } from './config/default-occ-opf-order-config';
 import { defaultOccOtpConfig } from './config/default-occ-otp-config';
 
@@ -48,11 +42,6 @@ import { defaultOccOtpConfig } from './config/default-occ-otp-config';
     {
       provide: OtpAdapter,
       useClass: OccOtpAdapter,
-    },
-    provideDefaultConfig(defaultOccOpfConfigurationConfig),
-    {
-      provide: OpfConfigurationAdapter,
-      useClass: OccOpfConfigurationAdapter,
     },
   ],
 })
