+++ integration-libs/opf/base/root/facade/index.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './opf-configuration.facade';
 export * from './opf-global-functions.facade';
 export * from './opf-order.facade';
 export * from './opf-otp.facade';
